boco-backend/requirements.txt
appdirs==1.4.3
coreapi==2.2.4
coreschema==0.0.4
coverage==4.3.4
Django==2.0.2
django-autoslug==1.9.3
django-debug-toolbar==1.6
django-fakery==1.9.1
django-cors-middleware==1.3.1
django-crispy-forms==1.6.1
django-filter==1.0.2
django-rest-swagger==2.1.1
django-recurrence==1.5.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-queryfields==1.0.0
dj-database-url==0.4.2
django-sslserver==0.19
django-url-filter==0.3.4
envparse==0.2.0
fake-factory==0.7.4
flake8==3.3.0
packaging==16.8
py==1.4.32
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.1.10
pytest==3.4.1
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-xdist==1.22.0
python-dateutil==2.6.0
python-json-logger==0.1.6
requests==2.13.0
simplejson==3.10.0
sqlparse==0.2.2
uritemplate==3.0.0
isort==4.2.15
django-simple-history==1.9.0
django-sequences==2.2
django-easy-pdf>=0.1.1<0.2.0
xhtml2pdf==0.2b1
pandas==0.20.3
django-pandas==0.4.5